+++ tests/check_command_allowed/test_command_not_allowed_ready.py
@@ -28,7 +28,6 @@ tmc_helper = TmcHelper(centralnode, tmc_subarraynode1)
 telescope_control = BaseTelescopeControl()
 
 
-@pytest.mark.skip(reason="Debugging")
 @pytest.mark.SKA_mid
 @scenario(
     "../features/check_command_not_allowed.feature",