+++ tests/tmc_csp/test_xtp_29345_configure.py
@@ -3,8 +3,7 @@ import json
 
 import pytest
 from pytest_bdd import given, parsers, scenario, then, when
-from ska_control_model import ObsState
-from tango import DevState
+from ska_control_model import ObsState, ResultCode
 
 from tests.resources.test_harness.central_node_mid import CentralNodeWrapperMid
 from tests.resources.test_harness.event_recorder import EventRecorder
@@ -14,10 +13,12 @@ from tests.resources.test_harness.helpers import (
     wait_till_delay_values_are_populated,
 )
 from tests.resources.test_harness.subarray_node import SubarrayNodeWrapper
-from tests.resources.test_harness.utils.common_utils import JsonFactory
+from tests.resources.test_harness.utils.common_utils import (
+    JsonFactory,
+    wait_added_for_skb372,
+)
 
 
-@pytest.mark.skip(reason="Requires CBF chart with ska-telmodel v.1.15.0")
 @pytest.mark.tmc_csp
 @scenario(
     "../features/tmc_csp/xtp_29345_configure.feature",
@@ -27,23 +28,6 @@ def test_tmc_csp_configure_functionality() -> None:
     """
     Test case to verify TMC-CSP Configure functionality
     """
-
-
-@given("the telescope is in ON state")
-def check_telescope_is_in_on_state(
-    central_node_mid: CentralNodeWrapperMid, event_recorder: EventRecorder
-) -> None:
-    """Ensure telescope is in ON state."""
-    central_node_mid.move_to_on()
-    event_recorder.subscribe_event(
-        central_node_mid.central_node, "telescopeState"
-    )
-    assert event_recorder.has_change_event_occurred(
-        central_node_mid.central_node,
-        "telescopeState",
-        DevState.ON,
-        lookahead=15,
-    )
 
 
 @given(parsers.parse("TMC subarray {subarray_id} in ObsState IDLE"))
@@ -61,17 +45,21 @@ def move_subarray_node_to_idle_obsstate(
     # Create json for AssignResources commands with requested subarray_id
     assign_input = json.loads(assign_input_json)
     assign_input["subarray_id"] = int(subarray_id)
-    central_node_mid.perform_action(
+    pytest.command_result = central_node_mid.perform_action(
         "AssignResources", json.dumps(assign_input)
     )
-
-    event_recorder.subscribe_event(central_node_mid.subarray_node, "obsState")
+    assert event_recorder.has_change_event_occurred(
+        central_node_mid.central_node,
+        "longRunningCommandResult",
+        (pytest.command_result[1][0], str(ResultCode.OK.value)),
+    )
     assert event_recorder.has_change_event_occurred(
         central_node_mid.subarray_node,
         "obsState",
         ObsState.IDLE,
         lookahead=20,
     )
+    wait_added_for_skb372()
 
 
 @when(
@@ -80,13 +68,22 @@ def move_subarray_node_to_idle_obsstate(
     )
 )
 def invoke_configure_command(
-    subarray_node: SubarrayNodeWrapper, command_input_factory: JsonFactory
+    subarray_node: SubarrayNodeWrapper,
+    command_input_factory: JsonFactory,
+    event_recorder: EventRecorder,
 ) -> None:
     """Invoke Configure command."""
     configure_input_json = prepare_json_args_for_commands(
         "configure_mid", command_input_factory
     )
-    subarray_node.execute_transition("Configure", argin=configure_input_json)
+    pytest.command_result = subarray_node.execute_transition(
+        "Configure", argin=configure_input_json
+    )
+    assert event_recorder.has_change_event_occurred(
+        subarray_node.subarray_node,
+        "longRunningCommandResult",
+        (pytest.command_result[1][0], str(ResultCode.OK.value)),
+    )
 
 
 @then(
@@ -98,9 +95,6 @@ def check_if_csp_subarray_moved_to_ready_obsstate(
     subarray_node: SubarrayNodeWrapper, event_recorder: EventRecorder
 ) -> None:
     """Ensure CSP subarray is moved to READY obsstate"""
-    event_recorder.subscribe_event(
-        subarray_node.subarray_devices["csp_subarray"], "obsState"
-    )
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_devices["csp_subarray"],
         "obsState",