+++ tests/tmc_csp/test_xtp_29374_end.py
@@ -5,7 +5,6 @@ import logging
 import pytest
 from pytest_bdd import given, parsers, scenario, then, when
 from ska_control_model import ObsState
-from tango import DevState
 
 from tests.resources.test_harness.central_node_mid import CentralNodeWrapperMid
 from tests.resources.test_harness.event_recorder import EventRecorder
@@ -19,7 +18,6 @@ from tests.resources.test_harness.utils.common_utils import JsonFactory
 LOGGER = logging.getLogger(__name__)
 
 
-@pytest.mark.skip(reason="Requires CBF chart with ska-telmodel v.1.15.0")
 @pytest.mark.tmc_csp
 @scenario(
     "../features/tmc_csp/xtp_29374_end.feature",
@@ -29,27 +27,6 @@ def test_tmc_csp_end_functionality() -> None:
     """
     Test case to verify TMC-CSP observation End functionality
     """
-
-
-@given("the telescope is in ON state")
-def check_telescope_is_in_on_state(
-    central_node_mid: CentralNodeWrapperMid, event_recorder: EventRecorder
-) -> None:
-    """Ensure telescope is in ON state."""
-    if central_node_mid.telescope_state != "ON":
-        central_node_mid.wait.set_wait_for_csp_master_to_become_off()
-        central_node_mid.csp_master.adminMode = 0
-        central_node_mid.wait.wait(500)
-        central_node_mid.move_to_on()
-    event_recorder.subscribe_event(
-        central_node_mid.central_node, "telescopeState"
-    )
-    assert event_recorder.has_change_event_occurred(
-        central_node_mid.central_node,
-        "telescopeState",
-        DevState.ON,
-        lookahead=15,
-    )
 
 
 @given(parsers.parse("TMC subarray {subarray_id} is in READY ObsState"))
@@ -72,7 +49,6 @@ def move_subarray_node_to_ready_obsstate(
         "AssignResources", json.dumps(assign_input)
     )
 
-    event_recorder.subscribe_event(central_node_mid.subarray_node, "obsState")
     assert event_recorder.has_change_event_occurred(
         central_node_mid.subarray_node,
         "obsState",
@@ -102,9 +78,6 @@ def check_if_csp_subarray_moved_to_idle_obsstate(
     subarray_node: SubarrayNodeWrapper, event_recorder
 ):
     """Ensure CSP subarray is moved to IDLE obsstate"""
-    event_recorder.subscribe_event(
-        subarray_node.subarray_devices["csp_subarray"], "obsState"
-    )
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_devices["csp_subarray"],
         "obsState",