+++ tests/tmc_csp/test_xtp_29738_restart.py
@@ -2,7 +2,6 @@
 import pytest
 from pytest_bdd import given, parsers, scenario, then, when
 from ska_control_model import ObsState
-from tango import DevState
 
 from tests.resources.test_harness.helpers import (
     prepare_json_args_for_centralnode_commands,
@@ -18,20 +17,6 @@ def test_tmc_csp_restart(central_node_mid):
     """
     Test case to verify TMC-CSP Restart functionality
     """
-
-
-@given("the telescope is in ON state")
-def telescope_is_in_on_state(central_node_mid, event_recorder):
-    """A method to check if telescope in is on state."""
-    central_node_mid.move_to_on()
-    event_recorder.subscribe_event(
-        central_node_mid.central_node, "telescopeState"
-    )
-    assert event_recorder.has_change_event_occurred(
-        central_node_mid.central_node,
-        "telescopeState",
-        DevState.ON,
-    )
 
 
 @given(
@@ -50,10 +35,6 @@ def subarray_is_in_aborted_obsstate(
         "assign_resources_mid", command_input_factory
     )
     central_node_mid.store_resources(assign_input_json)
-    event_recorder.subscribe_event(
-        subarray_node.subarray_devices.get("csp_subarray"), "obsState"
-    )
-    event_recorder.subscribe_event(subarray_node.subarray_node, "obsState")
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_devices.get("csp_subarray"),
         "obsState",