+++ tests/tmc_sdp/test_xtp-29382_configure.py
@@ -12,9 +12,6 @@ from tests.resources.test_harness.helpers import (
 )
 
 
-@pytest.mark.skip(
-    reason="getting Corba calltimeout issue in dishleafnode logs"
-)
 @pytest.mark.tmc_sdp
 @scenario(
     "../features/tmc_sdp/xtp-29382_configure.feature",
@@ -68,6 +65,11 @@ def check_subarray_obs_state(
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_devices.get("sdp_subarray"),
         "obsState",
+        ObsState.RESOURCING,
+    )
+    assert event_recorder.has_change_event_occurred(
+        subarray_node.subarray_devices.get("sdp_subarray"),
+        "obsState",
         ObsState.IDLE,
     )
     assert event_recorder.has_change_event_occurred(
@@ -103,6 +105,11 @@ def check_sdp_subarray_in_ready(
 ):
     """A method to check SDP subarray obsstate"""
     central_node_mid.set_subarray_id(subarray_id)
+    assert event_recorder.has_change_event_occurred(
+        subarray_node.subarray_devices["sdp_subarray"],
+        "obsState",
+        ObsState.CONFIGURING,
+    )
     assert event_recorder.has_change_event_occurred(
         subarray_node.subarray_devices["sdp_subarray"],
         "obsState",