+++ tests/tmc_sdp/test_xtp-32451_duplicate_ebid_exception.py
@@ -110,9 +110,7 @@ def given_assign_resources_executed_on_tmc_subarray(
         "AssignResources", assign_input_json
     )
     assert event_recorder.has_change_event_occurred(
-        subarray_node.subarray_node,
-        "obsState",
-        ObsState.IDLE,
+        subarray_node.subarray_node, "obsState", ObsState.IDLE, lookahead=15
     )
     assert event_recorder.has_change_event_occurred(
         central_node_mid.central_node,
@@ -331,6 +329,7 @@ def assign_resources_executed_on_subarray(
         subarray_node.subarray_devices.get("sdp_subarray"),
         "obsState",
         ObsState.IDLE,
+        lookahead=10,
     )
 
     check_subarray_instance(subarray_node.subarray_node, subarray_id)
@@ -338,4 +337,5 @@ def assign_resources_executed_on_subarray(
         subarray_node.subarray_node,
         "obsState",
         ObsState.IDLE,
+        lookahead=10,
     )